remove repository provider
