+++ src/docfx/lib/git/GitCommitProvider.cs
@@ -1,22 +1,28 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
+using System.IO;
 using System.Threading.Tasks;
 
 namespace Microsoft.Docs.Build
 {
     internal sealed class GitCommitProvider : IDisposable
     {
+        private readonly ConcurrentDictionary<string, Repository> _repositoryByFolder = new ConcurrentDictionary<string, Repository>();
         private readonly ConcurrentDictionary<string, Lazy<Task<(FileCommitProvider provider, Repository repo)>>> _commitProviders = new ConcurrentDictionary<string, Lazy<Task<(FileCommitProvider provider, Repository repo)>>>();
 
-        public async Task<List<GitCommit>> GetCommitHistory(Repository repo, string file, string committish = null)
-        {
-            var fileCommitProvider = await GetGitCommitProvider(repo);
+        public Task<(Repository repo, string pathToRepo, List<GitCommit> commits)> GetCommitHistory(Document document, string committish = null)
+           => GetCommitHistory(Path.Combine(document.Docset.DocsetPath, document.FilePath), committish);
 
-            return fileCommitProvider.GetCommitHistory(file, committish);
+        public async Task<(Repository repo, string pathToRepo, List<GitCommit> commits)> GetCommitHistory(string fullPath, string committish = null)
+        {
+            var repo = GetRepository(fullPath);
+            if (repo == null)
+                return default;
+            var pathToRepo = PathUtility.NormalizeFile(Path.GetRelativePath(repo.Path, fullPath));
+            return (repo, pathToRepo, await GetCommitHistory(repo, pathToRepo, committish));
         }
 
         public async Task SaveGitCommitCache()
@@ -36,6 +42,13 @@ namespace Microsoft.Docs.Build
             }
         }
 
+        private async Task<List<GitCommit>> GetCommitHistory(Repository repo, string file, string committish = null)
+        {
+            var fileCommitProvider = await GetGitCommitProvider(repo);
+
+            return fileCommitProvider.GetCommitHistory(file, committish);
+        }
+
         private async Task<FileCommitProvider> GetGitCommitProvider(Repository repo)
         {
             var (provider, _) = await _commitProviders.GetOrAdd(
@@ -46,6 +59,16 @@ namespace Microsoft.Docs.Build
                 })).Value;
 
             return provider;
+        }
+
+        private Repository GetRepository(string fullPath)
+        {
+            if (GitUtility.IsRepo(fullPath))
+                return Repository.Create(fullPath);
+            var parent = Path.GetDirectoryName(fullPath);
+            return !string.IsNullOrEmpty(parent)
+                ? _repositoryByFolder.GetOrAdd(parent, GetRepository)
+                : null;
         }
     }
 }
