+++ test/docfx.Test/lib/GitUtilityTest.cs
@@ -35,15 +35,13 @@ namespace Microsoft.Docs.Build
             var repoPath = GitUtility.FindRepo(Path.GetFullPath(file));
             Assert.NotNull(repoPath);
 
-            var repo = Repository.Create(repoPath);
-
-            using (var commitsProvider = new GitCommitProvider())
+            using (var gitCommitProvider = new GitCommitProvider())
             {
                 var pathToRepo = PathUtility.NormalizeFile(file);
 
                 // current branch
                 var exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" -- \"{pathToRepo}\"", repoPath);
-                var lib = await commitsProvider.GetCommitHistory(repo, pathToRepo);
+                var (_, _, lib) = await gitCommitProvider.GetCommitHistory(Path.Combine(repoPath, pathToRepo));
 
                 Assert.Equal(
                     exe.Replace("\r", ""),
@@ -51,13 +49,13 @@ namespace Microsoft.Docs.Build
 
                 // another branch
                 exe = Exec("git", $"--no-pager log --format=\"%H|%cI|%an|%ae\" a050eaf -- \"{pathToRepo}\"", repoPath);
-                lib = await commitsProvider.GetCommitHistory(repo, pathToRepo, "a050eaf");
+                (_, _, lib) = await gitCommitProvider.GetCommitHistory(Path.Combine(repoPath, pathToRepo), "a050eaf");
 
                 Assert.Equal(
                     exe.Replace("\r", ""),
                     string.Join("\n", lib.Select(c => $"{c.Sha}|{c.Time.ToString("s")}{c.Time.ToString("zzz")}|{c.AuthorName}|{c.AuthorEmail}")));
 
-                await commitsProvider.SaveGitCommitCache();
+                await gitCommitProvider.SaveGitCommitCache();
             }
 
         }
